Readme
Because of the high price of petrol, you are concerned with the fuel consumption of
your car. As a result you have a record of the kilometers driven and liters used for each tank of
petrol you purchase. Write a program that will display the kilometers driven, liters used, and
consumption (in litres/100km) for each tankful. After processing all the input information, the
program should calculate the overall average consumption: (Hint The liters/100km is (Liters/KM
driven) x 100)